#386 rename convergence tests to integration, move model testing in there as well
