+++ tests/__init__.py
@@ -2,8 +2,8 @@
 # Root of the tests module.
 # Provides access to all shared functionality
 #
-from .unit.test_models.standard_model_tests import StandardModelTest, OptimisationsTest
-from .unit.test_models.standard_output_tests import StandardOutputTests
+from .integration.test_models.standard_model_tests import StandardModelTest, OptimisationsTest
+from .integration.test_models.standard_output_tests import StandardOutputTests
 from .shared import (
     get_mesh_for_testing,
     get_p2d_mesh_for_testing,
